fixed compiler warning, added test

+++ BkpClasses/Arr.h
@@ -75,7 +75,7 @@ namespace bkp {
         // iterator constructor
         // Note: requires size() function on TContainer to work properly
         template<typename TContainer>
-        explicit Arr(const TContainer& srcCtnr) : Arr(srcCtnr.size(), srcCtnr.begin(), srcCtnr.end()) { }
+        explicit Arr(const TContainer& srcCtnr) : Arr(static_cast<int>(srcCtnr.size()), srcCtnr.begin(), srcCtnr.end()) { }
         
         // destructor
         virtual ~Arr() {
